fix(search): make search work and searchbad not
